feat: コネクションステートとNone型の定義およびzodスキーマを追加
