+++ packages/ws-signal/src/core/connection/connection-state.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+import { None, NoneSchema } from "./none";
+
+/** コネクションステート */
+export type ConnectionState = None;
+
+/** ConnectionState zodスキーマ */
+export const ConnectionStateSchema = z.union([NoneSchema]);
